// src/app/models/user-notification.ts
export class UserNotification {
    NotificationID: number;
    NotificationTypeID: number;
    TextData: string;
    RecipientID: number;
    RedirectURL: string;
    IsSent: boolean;
    IsRead: boolean;
    JobCardID: number;
    IsNew: boolean;
    TotalNotifications: number;
    UnReadNotifications: number;
}